requirements.txt
numpy==1.18.5
tensorflow==2.3.2
imageio==2.9.0
scikit-image==0.17.2
Pillow==8.1.2
scipy==1.7.1
scikit-learn==0.24.1
